+++ packages/blocks/src/api/registration.js
@@ -146,6 +146,7 @@ export function registerBlockType( name, settings ) {
 		attributes: {},
 		keywords: [],
 		save: () => null,
+		allowDefaultStyle: true,
 		...get( serverSideBlockDefinitions, name ),
 		...settings,
 	};