+++ packages/blocks/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Master
+
+### New Feature
+
+- Blocks can now be registered with an `allowDefaultStyle` setting, allowing the default style picker to be removed.
+
 ## 6.4.0 (2019-08-05)
 
 ### Improvements